pyproject.toml
[project]
name = "sea-guard"
version = "0.1.0"
description = "GuardBench — open-source AI safety guard evaluation framework"
requires-python = ">=3.9"
dependencies = [
    "click>=8.0",
    "jinja2>=3.0",
    "pyyaml>=6.0",
    "pandas>=1.5",
    "scikit-learn>=1.1",
    "matplotlib>=3.5",
    "pydantic>=2.0",
]

[project.optional-dependencies]
llm = ["openai", "anthropic"]
hf = ["datasets", "transformers"]
pdf = ["weasyprint"]
dev = [
    "pytest>=7.0",
    "pytest-cov>=4.0",
    "ruff>=0.1.0",
    "mypy>=1.0",
]

[project.scripts]
guardbench = "guardbench.cli.main:cli"

[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]
include = ["guardbench*"]

[tool.setuptools.package-data]
"guardbench.report.templates" = ["*.html"]
"guardbench.data.builtin" = ["*.csv", "*.jsonl"]

[tool.pytest.ini_options]
testpaths = ["tests/guardbench"]
addopts = "--tb=short"
